Extend tool support for queries
Provide both the namespace and the associated cmd line when queried about namespaces. Track the cmd line for tools and which tools are classified as launchers.
